+++ services/cli/src/actions/identity-x/create-behavior-categories.js
