+++ services/cli/src/actions/identity-x/create-behaviors.js
